+++ samples/test/lightweight_python_functions_v2_pipeline.py
@@ -15,10 +15,8 @@
 from typing import Dict, List
 import os
 
-from kfp import dsl
-from kfp import components
-from kfp.components import InputPath, OutputPath
-from kfp.v2.dsl import Input, Output, Dataset, Model, component
+from kfp.v2 import dsl
+from kfp.v2.dsl import Input, InputPath, Output, OutputPath, Dataset, Model, component
 import kfp.v2.compiler as compiler
 
 # In tests, we install a KFP package from the PR under test. Users should not
@@ -126,12 +124,12 @@ def train(
 def pipeline(message: str = 'message'):
     preprocess_task = preprocess(message=message)
     train_task = train(
-        dataset_one=preprocess_task.outputs['output_dataset_one'],
-        dataset_two=preprocess_task.outputs['output_dataset_two'],
-        message=preprocess_task.outputs['output_parameter'],
-        input_bool=preprocess_task.outputs['output_bool_parameter'],
-        input_dict=preprocess_task.outputs['output_dict_parameter'],
-        input_list=preprocess_task.outputs['output_list_parameter'],
+        dataset_one_path=preprocess_task.outputs['output_dataset_one'],
+        dataset_two=preprocess_task.outputs['output_dataset_two_path'],
+        message=preprocess_task.outputs['output_parameter_path'],
+        input_bool=preprocess_task.outputs['output_bool_parameter_path'],
+        input_dict=preprocess_task.outputs['output_dict_parameter_path'],
+        input_list=preprocess_task.outputs['output_list_parameter_path'],
     )
