+++ sdk/python/kfp/v2/compiler/compiler_test.py
@@ -18,7 +18,7 @@ import shutil
 import tempfile
 import unittest
 
-from kfp.v2 import components
+from kfp import components
 from kfp.v2 import compiler
 from kfp.v2 import dsl
 from kfp.dsl import types
