+++ sdk/python/kfp/v2/components/types/test_data/expected_confusion_matrix.json
@@ -1,0 +1,10 @@
+{
+  "annotationSpecs": [
+    {"displayName": "dog"},
+    {"displayName": "cat"},
+    {"displayName": "horses"}],
+  "row": [
+    [2, 6, 0],
+    [3, 0, 0],
+    [0, 0, 0]]
+}
