+++ sdk/python/kfp/v2/components/types/test_data/expected_io_types_bulk_load_classification_metrics.json
