+++ sdk/python/kfp/v2/components/types/type_utils.py
@@ -17,18 +17,17 @@ from typing import Dict, List, Optional, Type, Union
 from kfp.components import structures
 from kfp.components import type_annotation_utils
 from kfp.pipeline_spec import pipeline_spec_pb2
-from kfp.dsl import artifact_utils
-from kfp.dsl import io_types
+from kfp.v2.components.types import artifact_types
 
 # ComponentSpec I/O types to DSL ontology artifact classes mapping.
 _ARTIFACT_CLASSES_MAPPING = {
-    'model': io_types.Model,
-    'dataset': io_types.Dataset,
-    'metrics': io_types.Metrics,
-    'classificationmetrics': io_types.ClassificationMetrics,
-    'slicedclassificationmetrics': io_types.SlicedClassificationMetrics,
-    'html': io_types.HTML,
-    'markdown': io_types.Markdown,
+    'model': artifact_types.Model,
+    'dataset': artifact_types.Dataset,
+    'metrics': artifact_types.Metrics,
+    'classificationmetrics': artifact_types.ClassificationMetrics,
+    'slicedclassificationmetrics': artifact_types.SlicedClassificationMetrics,
+    'html': artifact_types.HTML,
+    'markdown': artifact_types.Markdown,
 }
 
 # ComponentSpec I/O types to (IR) PipelineTaskSpec I/O types mapping.
@@ -60,7 +59,7 @@ _PARAMETER_TYPES_VALUE_REFERENCE_MAPPING = {
 
 
 def is_parameter_type(type_name: Optional[Union[str, dict]]) -> bool:
-  """Check if a ComponentSpec I/O type is considered as a parameter type.
+    """Check if a ComponentSpec I/O type is considered as a parameter type.
 
   Args:
     type_name: type name of the ComponentSpec I/O type.
@@ -68,36 +67,37 @@ def is_parameter_type(type_name: Optional[Union[str, dict]]) -> bool:
   Returns:
     True if the type name maps to a parameter type else False.
   """
-  if isinstance(type_name, str):
-    type_name = type_annotation_utils.get_short_type_name(type_name)
-  elif isinstance(type_name, dict):
-    type_name = list(type_name.keys())[0]
-  else:
-    return False
+    if isinstance(type_name, str):
+        type_name = type_annotation_utils.get_short_type_name(type_name)
+    elif isinstance(type_name, dict):
+        type_name = list(type_name.keys())[0]
+    else:
+        return False
 
-  return type_name.lower() in _PARAMETER_TYPES_MAPPING
+    return type_name.lower() in _PARAMETER_TYPES_MAPPING
 
 
 def get_artifact_type_schema(
-    artifact_class_or_type_name: Optional[Union[str, Type[io_types.Artifact]]]
+    artifact_class_or_type_name: Optional[Union[str,
+                                                Type[artifact_types.Artifact]]]
 ) -> pipeline_spec_pb2.ArtifactTypeSchema:
-  """Gets the IR I/O artifact type msg for the given ComponentSpec I/O type."""
-  artifact_class = io_types.Artifact
-  if isinstance(artifact_class_or_type_name, str):
-    artifact_class = _ARTIFACT_CLASSES_MAPPING.get(
-        artifact_class_or_type_name.lower(), io_types.Artifact)
-  elif inspect.isclass(artifact_class_or_type_name) and issubclass(
-      artifact_class_or_type_name, io_types.Artifact):
-    artifact_class = artifact_class_or_type_name
+    """Gets the IR I/O artifact type msg for the given ComponentSpec I/O type."""
+    artifact_class = artifact_types.Artifact
+    if isinstance(artifact_class_or_type_name, str):
+        artifact_class = _ARTIFACT_CLASSES_MAPPING.get(
+            artifact_class_or_type_name.lower(), artifact_types.Artifact)
+    elif inspect.isclass(artifact_class_or_type_name) and issubclass(
+            artifact_class_or_type_name, artifact_types.Artifact):
+        artifact_class = artifact_class_or_type_name
 
-  return pipeline_spec_pb2.ArtifactTypeSchema(
-      schema_title=artifact_class.TYPE_NAME)
+    return pipeline_spec_pb2.ArtifactTypeSchema(
+        schema_title=artifact_class.TYPE_NAME)
 
 
 def get_parameter_type(
     param_type: Optional[Union[Type, str, dict]]
 ) -> pipeline_spec_pb2.PrimitiveType:
-  """Get the IR I/O parameter type for the given ComponentSpec I/O type.
+    """Get the IR I/O parameter type for the given ComponentSpec I/O type.
 
   Args:
     param_type: type of the ComponentSpec I/O type. Can be a primitive Python
@@ -109,17 +109,17 @@ def get_parameter_type(
   Raises:
     AttributeError: if type_name is not a string type.
   """
-  if type(param_type) == type:
-    type_name = param_type.__name__
-  elif isinstance(param_type, dict):
-    type_name = list(param_type.keys())[0]
-  else:
-    type_name = type_annotation_utils.get_short_type_name(str(param_type))
-  return _PARAMETER_TYPES_MAPPING.get(type_name.lower())
+    if type(param_type) == type:
+        type_name = param_type.__name__
+    elif isinstance(param_type, dict):
+        type_name = list(param_type.keys())[0]
+    else:
+        type_name = type_annotation_utils.get_short_type_name(str(param_type))
+    return _PARAMETER_TYPES_MAPPING.get(type_name.lower())
 
 
 def get_parameter_type_field_name(type_name: Optional[str]) -> str:
-  """Get the IR field name for the given primitive type.
+    """Get the IR field name for the given primitive type.
 
   For example: 'str' -> 'string_value', 'double' -> 'double_value', etc.
 
@@ -132,15 +132,15 @@ def get_parameter_type_field_name(type_name: Optional[str]) -> str:
   Raises:
     AttributeError: if type_name is not a string type.
   """
-  return _PARAMETER_TYPES_VALUE_REFERENCE_MAPPING.get(
-      get_parameter_type(type_name))
+    return _PARAMETER_TYPES_VALUE_REFERENCE_MAPPING.get(
+        get_parameter_type(type_name))
 
 
 def get_input_artifact_type_schema(
     input_name: str,
     inputs: List[structures.InputSpec],
 ) -> Optional[str]:
-  """Find the input artifact type by input name.
+    """Find the input artifact type by input name.
 
   Args:
     input_name: The name of the component input.
@@ -152,9 +152,9 @@ def get_input_artifact_type_schema(
   Raises:
     AssertionError if input not found, or input found but not an artifact type.
   """
-  for component_input in inputs:
-    if component_input.name == input_name:
-      assert not is_parameter_type(
-          component_input.type), 'Input is not an artifact type.'
-      return get_artifact_type_schema(component_input.type)
-  assert False, 'Input not found.'
+    for component_input in inputs:
+        if component_input.name == input_name:
+            assert not is_parameter_type(
+                component_input.type), 'Input is not an artifact type.'
+            return get_artifact_type_schema(component_input.type)
+    assert False, 'Input not found.'
