+++ sdk/python/kfp/v2/dsl/__init__.py
@@ -12,23 +12,28 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from kfp.v2.dsl.component_decorator import component
-from kfp.dsl.io_types import (
-    Input,
-    Output,
+from kfp.v2.components.component_decorator import component
+
+from kfp.v2.components.importer_node import importer
+
+from kfp.v2.components.types.artifact_types import (
     Artifact,
-    Dataset,
-    Model,
-    Metrics,
     ClassificationMetrics,
-    SlicedClassificationMetrics,
+    Dataset,
     HTML,
     Markdown,
+    Metrics,
+    Model,
+    SlicedClassificationMetrics,
 )
-from kfp.components import (
+
+from kfp.v2.components.types.type_annotations import (
+    Input,
+    Output,
     InputPath,
     OutputPath,
 )
+
 from kfp.dsl import (
     graph_component,
     pipeline,
